Squashed commit of the following:
commit 778d130e2069f671a1ded90fa541c001aac68003
Author: Martin Poulton <mapoulton@gmail.com>
Date:   Wed Jan 19 18:53:01 2022 +0000

    Refactor of class locations

commit a33da8ca407848bfc86a2f8d8f35c7af3cbad806
Author: Martin Poulton <mapoulton@gmail.com>
Date:   Wed Jan 19 18:50:47 2022 +0000

    FindUsersWIthinDistanceOf_ReturnsResults - Refactor

commit 304b3fa2d707cf1080fbaaece32e3e1a7bb1eb3e
Author: Martin Poulton <mapoulton@gmail.com>
Date:   Wed Jan 19 18:39:32 2022 +0000

    FindUsersWIthinDistanceOf_ReturnsResults - Passing Test

commit eb8a2eeff86995da7aae016658c5747bea6015c5
Author: Martin Poulton <mapoulton@gmail.com>
Date:   Wed Jan 19 18:32:19 2022 +0000

    FindUsersWIthinDistanceOf_ReturnsResults - Failing test

commit ab8210e88eb153f0a8a176d442acc12492790748
Author: Martin Poulton <mapoulton@gmail.com>
Date:   Wed Jan 19 15:23:00 2022 +0000

    Added Distance Calculator

+++ src/DWP.Demo.Api/Types/User.cs
@@ -7,7 +7,7 @@ namespace DWP.Demo.Api.Types
         public string LastName { get; init; }
         public string Email { get; init; }
         public string IpAddress { get; init; }
-        public decimal Latitude { get; init; }
-        public decimal Longitude { get; init; }
+        public double Latitude { get; init; }
+        public double Longitude { get; init; }
     }
 }
